Refactor: `core_wrapper` 移動
