+++ voicevox_engine/core_wrapper.py
