+++ voicevox_engine/tts_pipeline/make_synthesis_engines.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import Dict, List, Optional
 
 from ..utility import engine_root, get_save_dir
-from .core_wrapper import CoreWrapper, load_runtime_lib
+from ..core_wrapper import CoreWrapper, load_runtime_lib
 from .tts_engine import SynthesisEngine, SynthesisEngineBase
