+++ voicevox_engine/tts_pipeline/tts_engine.py
@@ -8,7 +8,7 @@ from soxr import resample
 
 from .acoustic_feature_extractor import OjtPhoneme
 from ..model import AccentPhrase, AudioQuery, Mora
-from .core_wrapper import CoreWrapper, OldCoreError
+from ..core_wrapper import CoreWrapper, OldCoreError
 from .tts_engine_base import SynthesisEngineBase
 
 unvoiced_mora_phoneme_list = ["A", "I", "U", "E", "O", "cl", "pau"]
